Rename folder for custom testing platform targets
