+++ packages/constraint-engine/src/customRules/packageAHasTypeScriptConfigFile.ts
@@ -2,7 +2,7 @@ import { ParseableOnDiskJsonFileTarget } from '../customTargets/file/jsonFileTar
 import {
   ObjectTarget,
   TestingPlatformPackageATarget,
-} from '../customTargets/testingPlatformPackage/targets';
+} from '../customTargets/testingPlatform/targets';
 import { GuardRule } from '../types/rule';
 import { isObject } from './isObject';
